move tests to more appropriate locations
