+++ Assets/Tests/Runtime/NetworkServerTests.cs
@@ -3,24 +3,13 @@ using System.Collections;
 using Cysharp.Threading.Tasks;
 using NSubstitute;
 using NUnit.Framework;
-using UnityEngine;
-using UnityEngine.Events;
 using UnityEngine.TestTools;
 
 namespace Mirror.Tests
 {
-    public class SimpleNetworkServer : ServerObjectManager
-    {
-        public void SpawnObjectExpose(GameObject obj, INetworkConnection ownerConnection)
-        {
-            SpawnObject(obj, ownerConnection);
-        }
-    }
-
     public class NetworkServerTests : ClientServerSetup<MockComponent>
     {
         WovenTestMessage message;
-        GameObject playerReplacement;
 
         public override void ExtraSetup()
         {
@@ -44,53 +33,10 @@ namespace Mirror.Tests
         [Test]
         public void SendToClientOfPlayerExceptionTest()
         {
-            SimpleNetworkServer comp = serverPlayerGO.AddComponent<SimpleNetworkServer>();
-
             Assert.Throws<InvalidOperationException>(() =>
             {
                 server.SendToClientOfPlayer<ServerRpcMessage>(null, new ServerRpcMessage());
             });
-        }
-
-        [Test]
-        public void SpawnObjectExposeExceptionTest()
-        {
-            var gameObject = new GameObject();
-            SimpleNetworkServer comp = gameObject.AddComponent<SimpleNetworkServer>();
-
-            var obj = new GameObject();
-
-            InvalidOperationException ex = Assert.Throws<InvalidOperationException>(() =>
-            {
-                comp.SpawnObjectExpose(obj, connectionToServer);
-            });
-
-            Assert.That(ex.Message, Is.EqualTo("SpawnObject for " + obj + ", NetworkServer is not active. Cannot spawn objects without an active server."));
-        }
-
-        [Test]
-        public void SpawnNoIdentExceptionTest()
-        {
-            InvalidOperationException ex = Assert.Throws<InvalidOperationException>(() =>
-            {
-                serverObjectManager.Spawn(new GameObject(), new GameObject());
-            });
-
-            Assert.That(ex.Message, Is.EqualTo("Player object has no NetworkIdentity"));
-        }
-
-        [Test]
-        public void SpawnNotPlayerExceptionTest()
-        {
-            var player = new GameObject();
-            player.AddComponent<NetworkIdentity>();
-
-            InvalidOperationException ex = Assert.Throws<InvalidOperationException>(() =>
-            {
-                serverObjectManager.Spawn(new GameObject(), player);
-            });
-
-            Assert.That(ex.Message, Is.EqualTo("Player object is not a player in the connection"));
         }
 
         [UnityTest]
@@ -133,43 +79,6 @@ namespace Mirror.Tests
         });
 
         [UnityTest]
-        public IEnumerator ShowForConnection() => UniTask.ToCoroutine(async () =>
-        {
-            bool invoked = false;
-
-            connectionToServer.RegisterHandler<SpawnMessage>(msg => invoked = true) ;
-
-            connectionToClient.IsReady = true;
-
-            // call ShowForConnection
-            serverObjectManager.ShowForConnection(serverIdentity, connectionToClient);
-
-            connectionToServer.ProcessMessagesAsync().Forget();
-
-            await AsyncUtil.WaitUntilWithTimeout(() => invoked);
-        });
-
-        [Test]
-        public void SpawnSceneObject()
-        {
-            serverIdentity.sceneId = 42;
-            // unspawned scene objects are set to inactive before spawning
-            serverIdentity.gameObject.SetActive(false);
-            Assert.That(serverObjectManager.SpawnObjects(), Is.True);
-            Assert.That(serverIdentity.gameObject.activeSelf, Is.True);
-        }
-
-        [Test]
-        public void SpawnPrefabObject()
-        {
-            serverIdentity.sceneId = 0;
-            // unspawned scene objects are set to inactive before spawning
-            serverIdentity.gameObject.SetActive(false);
-            Assert.That(serverObjectManager.SpawnObjects(), Is.True);
-            Assert.That(serverIdentity.gameObject.activeSelf, Is.False);
-        }
-
-        [UnityTest]
         public IEnumerator RegisterMessage1() => UniTask.ToCoroutine(async () =>
         {
             bool invoked = false;
@@ -210,81 +119,9 @@ namespace Mirror.Tests
         });
 
         [Test]
-        public void ServerChangeSceneTest()
-        {
-            UnityAction<string, SceneOperation> func1 = Substitute.For<UnityAction<string, SceneOperation>>();
-            serverSceneManager.ServerChangeScene.AddListener(func1);
-            serverSceneManager.OnServerChangeScene("test", SceneOperation.Normal);
-            func1.Received(1).Invoke(Arg.Any<string>(), Arg.Any<SceneOperation>());
-        }
-
-        [Test]
-        public void ServerSceneChangedTest()
-        {
-            UnityAction<string, SceneOperation> func1 = Substitute.For<UnityAction<string, SceneOperation>>();
-            serverSceneManager.ServerSceneChanged.AddListener(func1);
-            serverSceneManager.OnServerSceneChanged("test", SceneOperation.Normal);
-            func1.Received(1).Invoke(Arg.Any<string>(), Arg.Any<SceneOperation>());
-        }
-
-        [Test]
-        public void ReplacePlayerBaseTest()
-        {
-            playerReplacement = new GameObject("replacement", typeof(NetworkIdentity));
-            NetworkIdentity replacementIdentity = playerReplacement.GetComponent<NetworkIdentity>();
-            replacementIdentity.AssetId = Guid.NewGuid();
-            clientObjectManager.RegisterPrefab(playerReplacement);
-
-            serverObjectManager.ReplacePlayerForConnection(connectionToClient, client, playerReplacement);
-
-            Assert.That(connectionToClient.Identity, Is.EqualTo(replacementIdentity));
-        }
-
-        [Test]
-        public void ReplacePlayerDontKeepAuthTest()
-        {
-            playerReplacement = new GameObject("replacement", typeof(NetworkIdentity));
-            NetworkIdentity replacementIdentity = playerReplacement.GetComponent<NetworkIdentity>();
-            replacementIdentity.AssetId = Guid.NewGuid();
-            clientObjectManager.RegisterPrefab(playerReplacement);
-
-            serverObjectManager.ReplacePlayerForConnection(connectionToClient, client, playerReplacement, true);
-
-            Assert.That(clientIdentity.ConnectionToClient, Is.EqualTo(null));
-        }
-
-        [Test]
-        public void ReplacePlayerAssetIdTest()
-        {
-            var replacementGuid = Guid.NewGuid();
-            playerReplacement = new GameObject("replacement", typeof(NetworkIdentity));
-            NetworkIdentity replacementIdentity = playerReplacement.GetComponent<NetworkIdentity>();
-            replacementIdentity.AssetId = replacementGuid;
-            clientObjectManager.RegisterPrefab(playerReplacement);
-
-            serverObjectManager.ReplacePlayerForConnection(connectionToClient, client, playerReplacement, replacementGuid);
-
-            Assert.That(connectionToClient.Identity.AssetId, Is.EqualTo(replacementGuid));
-        }
-
-        [Test]
         public void NumPlayersTest()
         {
             Assert.That(server.NumPlayers, Is.EqualTo(1));
-        }
-
-        [Test]
-        public void AddPlayerForConnectionAssetIdTest()
-        {
-            var replacementGuid = Guid.NewGuid();
-            playerReplacement = new GameObject("replacement", typeof(NetworkIdentity));
-            NetworkIdentity replacementIdentity = playerReplacement.GetComponent<NetworkIdentity>();
-            replacementIdentity.AssetId = replacementGuid;
-            clientObjectManager.RegisterPrefab(playerReplacement);
-
-            connectionToClient.Identity = null;
-
-            Assert.That(serverObjectManager.AddPlayerForConnection(connectionToClient, playerReplacement, replacementGuid), Is.True);
         }
 
         [Test]
