+++ Assets/Tests/Runtime/ServerObjectManagerTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c0dd8b87ba37fa4bb64414b61690f0b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
